pyproject.toml
[project]
name = "nhp-immunogenomics"
version = "1.0.0"
description = "NHP immunogenomics resources from IPD-MHC and IPD-NHKIR databases"
requires-python = ">=3.12"
license = {text = "MIT"}
dependencies = [
    "biopython>=1.84",
    "requests>=2.31",
]

[project.optional-dependencies]
dev = [
    "pytest>=8.0",
    "pytest-cov>=5.0",
    "responses>=0.25",
]

[project.scripts]
nhp-pipeline = "pipeline.cli:main"

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[tool.hatch.build.targets.wheel]
packages = ["pipeline"]

[tool.pytest.ini_options]
testpaths = ["tests"]
addopts = "-v"
